pyproject.toml
[project]
name = "alpha"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "DanielBerns", email = "daniel.w.berns@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1.6",
    "markdown>=3.10.2",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.3",
]

[project.scripts]
alpha = "alpha:main"

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"
